+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements/calculated_measurements/database_definitions.py
