+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements/calculated_measurements/schema.py
