+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements/repository.py
@@ -1,8 +1,10 @@
 from pyspark.sql import SparkSession
 
-from opengeh_electrical_heating.infrastructure import CalculatedMeasurements
-from opengeh_electrical_heating.infrastructure.measurements.measurements_calculated.database_definitions import (
+from opengeh_electrical_heating.infrastructure.measurements.calculated_measurements.database_definitions import (
     CalculatedMeasurementsDatabase,
+)
+from opengeh_electrical_heating.infrastructure.measurements.calculated_measurements.wrapper import (
+    CalculatedMeasurements,
 )
 from opengeh_electrical_heating.infrastructure.measurements.measurements_gold.database_definitions import (
     MeasurementsGoldDatabase,
